requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0
anthropic>=0.7.0
celery>=5.3.0
redis>=5.0.0
Pillow>=10.2.0
